Redact stream refactoring
